Create response classes and scope the models under response since tightly coupled
