--- lib/yelp/models/coordinate.rb
@@ -1,7 +1,0 @@
-require 'yelp/models/base'
-
-module Model
-  class Coordinate < Base
-    attr_accessor :latitude, :longitude
-  end
-end
