--- lib/yelp/models/gift_certificate_option.rb
@@ -1,7 +1,0 @@
-require 'yelp/models/base'
-
-module Model
-  class GiftCertificateOption < Base
-    attr_accessor :formatted_price, :price
-  end
-end
