--- lib/yelp/models/rating.rb
@@ -1,7 +1,0 @@
-require 'yelp/models/base'
-
-module Model
-  class Rating < Base
-    attr_accessor :rating, :img_url
-  end
-end
