--- lib/yelp/models/review.rb
@@ -1,14 +1,0 @@
-require 'yelp/models/base'
-require 'yelp/models/rating'
-
-module Model
-  class Review < Base
-    attr_accessor :excerpt, :id, :rating, :time_created, :user
-
-    def initialize(json)
-      super(json)
-
-      @rating = parse(@rating, Rating)
-    end
-  end
-end
