--- lib/yelp/models/user.rb
@@ -1,7 +1,0 @@
-require 'yelp/models/base'
-
-module Model
-  class User < Base
-    attr_accessor :id, :image_url, :name
-  end
-end
