+++ lib/yelp/responses/business.rb
@@ -1,0 +1,6 @@
+module Yelp
+  module Response
+    class Business
+    end
+  end
+end
