+++ lib/yelp/responses/models/coordinate.rb
@@ -1,0 +1,11 @@
+require 'yelp/responses/base'
+
+module Yelp
+  module Response
+    module Model
+      class Coordinate < Response::Base
+        attr_accessor :latitude, :longitude
+      end
+    end
+  end
+end
