+++ lib/yelp/responses/phone_search.rb
@@ -1,0 +1,6 @@
+module Yelp
+  module Response
+    class PhoneSearch
+    end
+  end
+end
